+++ tokio-io/Cargo.toml
@@ -25,7 +25,10 @@ publish = false
 bytes = "0.4.7"
 log = "0.4"
 futures-core-preview = "0.3.0-alpha.17"
+memchr = "2.2"
 
 [dev-dependencies]
 pin-utils = "0.1.0-alpha.4"
+tokio = { version = "0.2.0", path = "../tokio" }
+futures-util-preview = "0.3.0-alpha.17"
 tokio-test = { version = "0.2.0", path = "../tokio-test" }
