+++ tokio-io/src/io/copy.rs
@@ -1,9 +1,9 @@
+use crate::{AsyncRead, AsyncWrite};
 use futures_core::ready;
 use std::future::Future;
 use std::io;
 use std::pin::Pin;
 use std::task::{Context, Poll};
-use tokio_io::{AsyncRead, AsyncWrite};
 
 #[derive(Debug)]
 #[must_use = "futures do nothing unless you `.await` or poll them"]
