+++ tokio-io/src/io/flush.rs
@@ -1,9 +1,8 @@
+use crate::AsyncWrite;
 use std::future::Future;
 use std::io;
 use std::pin::Pin;
 use std::task::{Context, Poll};
-
-use tokio_io::AsyncWrite;
 
 /// A future used to fully flush an I/O object.
 ///
