+++ tokio-io/src/io/read.rs
@@ -1,9 +1,9 @@
+use crate::AsyncRead;
 use std::future::Future;
 use std::io;
 use std::marker::Unpin;
 use std::pin::Pin;
 use std::task::{Context, Poll};
-use tokio_io::AsyncRead;
 
 /// Tries to read some bytes directly into the given `buf` in asynchronous
 /// manner, returning a future type.
