+++ tokio-io/src/io/read_to_end.rs
@@ -1,5 +1,4 @@
-use tokio_io::AsyncRead;
-
+use crate::AsyncRead;
 use futures_core::ready;
 use std::future::Future;
 use std::io;
