+++ tokio-io/src/io/read_to_string.rs
@@ -1,45 +1,44 @@
-use super::read_until::read_until_internal;
+use super::read_to_end::read_to_end_internal;
+use crate::AsyncRead;
 use futures_core::ready;
 use std::future::Future;
-use std::io;
-use std::mem;
 use std::pin::Pin;
-use std::str;
 use std::task::{Context, Poll};
-use tokio_io::AsyncBufRead;
+use std::{io, mem, str};
 
-/// Future for the [`read_line`](crate::io::AsyncBufReadExt::read_line) method.
+/// Future for the [`read_to_string`](super::AsyncReadExt::read_to_string) method.
 #[derive(Debug)]
 #[must_use = "futures do nothing unless you `.await` or poll them"]
-pub struct ReadLine<'a, R: ?Sized + Unpin> {
+pub struct ReadToString<'a, R: ?Sized + Unpin> {
     reader: &'a mut R,
     buf: &'a mut String,
     bytes: Vec<u8>,
-    read: usize,
+    start_len: usize,
 }
 
-impl<R: ?Sized + Unpin> Unpin for ReadLine<'_, R> {}
+impl<R: ?Sized + Unpin> Unpin for ReadToString<'_, R> {}
 
-pub(crate) fn read_line<'a, R>(reader: &'a mut R, buf: &'a mut String) -> ReadLine<'a, R>
+pub(crate) fn read_to_string<'a, R>(reader: &'a mut R, buf: &'a mut String) -> ReadToString<'a, R>
 where
-    R: AsyncBufRead + ?Sized + Unpin,
+    R: AsyncRead + ?Sized + Unpin,
 {
-    ReadLine {
+    let start_len = buf.len();
+    ReadToString {
         reader,
         bytes: unsafe { mem::replace(buf.as_mut_vec(), Vec::new()) },
         buf,
-        read: 0,
+        start_len,
     }
 }
 
-pub(super) fn read_line_internal<R: AsyncBufRead + ?Sized>(
+fn read_to_string_internal<R: AsyncRead + ?Sized>(
     reader: Pin<&mut R>,
     cx: &mut Context<'_>,
     buf: &mut String,
     bytes: &mut Vec<u8>,
-    read: &mut usize,
+    start_len: usize,
 ) -> Poll<io::Result<usize>> {
-    let ret = ready!(read_until_internal(reader, cx, b'\n', bytes, read));
+    let ret = ready!(read_to_end_internal(reader, cx, bytes, start_len));
     if str::from_utf8(&bytes).is_err() {
         Poll::Ready(ret.and_then(|_| {
             Err(io::Error::new(
@@ -49,14 +48,16 @@ pub(super) fn read_line_internal<R: AsyncBufRead + ?Sized>(
         }))
     } else {
         debug_assert!(buf.is_empty());
-        debug_assert_eq!(*read, 0);
         // Safety: `bytes` is a valid UTF-8 because `str::from_utf8` returned `Ok`.
         mem::swap(unsafe { buf.as_mut_vec() }, bytes);
         Poll::Ready(ret)
     }
 }
 
-impl<R: AsyncBufRead + ?Sized + Unpin> Future for ReadLine<'_, R> {
+impl<A> Future for ReadToString<'_, A>
+where
+    A: AsyncRead + ?Sized + Unpin,
+{
     type Output = io::Result<usize>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
@@ -64,8 +65,8 @@ impl<R: AsyncBufRead + ?Sized + Unpin> Future for ReadLine<'_, R> {
             reader,
             buf,
             bytes,
-            read,
+            start_len,
         } = &mut *self;
-        read_line_internal(Pin::new(reader), cx, buf, bytes, read)
+        read_to_string_internal(Pin::new(reader), cx, buf, bytes, *start_len)
     }
 }
