+++ tokio-io/src/io/write.rs
@@ -1,8 +1,8 @@
+use crate::AsyncWrite;
 use std::future::Future;
 use std::io;
 use std::pin::Pin;
 use std::task::{Context, Poll};
-use tokio_io::AsyncWrite;
 
 /// A future to write some of the buffer to an `AsyncWrite`.
 #[derive(Debug)]
