+++ tokio-io/tests/read_to_end.rs
@@ -1,17 +1,17 @@
 #![deny(warnings, rust_2018_idioms)]
 #![feature(async_await)]
 
-use tokio::io::{AsyncRead, AsyncReadExt};
+use tokio_io::{AsyncRead, AsyncReadExt};
 use tokio_test::assert_ok;
 
-use std::io;
 use std::pin::Pin;
 use std::task::{Context, Poll};
+use std::{cmp, io};
 
 #[tokio::test]
-async fn read_exact() {
+async fn read_to_end() {
     struct Rd {
-        val: &'static [u8; 11],
+        val: &'static [u8],
     }
 
     impl AsyncRead for Rd {
@@ -21,19 +21,20 @@ async fn read_exact() {
             buf: &mut [u8],
         ) -> Poll<io::Result<usize>> {
             let me = &mut *self;
-            let len = buf.len();
+            let len = cmp::min(buf.len(), me.val.len());
 
-            buf[..].copy_from_slice(&me.val[..len]);
-            Poll::Ready(Ok(buf.len()))
+            buf[..len].copy_from_slice(&me.val[..len]);
+            me.val = &me.val[len..];
+            Poll::Ready(Ok(len))
         }
     }
 
-    let mut buf = Box::new([0; 8]);
+    let mut buf = vec![];
     let mut rd = Rd {
         val: b"hello world",
     };
 
-    let n = assert_ok!(rd.read_exact(&mut buf[..]).await);
-    assert_eq!(n, 8);
-    assert_eq!(buf[..], b"hello wo"[..]);
+    let n = assert_ok!(rd.read_to_end(&mut buf).await);
+    assert_eq!(n, 11);
+    assert_eq!(buf[..], b"hello world"[..]);
 }
