+++ tokio-io/tests/write.rs
@@ -1,17 +1,16 @@
 #![deny(warnings, rust_2018_idioms)]
 #![feature(async_await)]
 
-use tokio::io::{AsyncWrite, AsyncWriteExt};
+use tokio_io::{AsyncWrite, AsyncWriteExt};
 use tokio_test::assert_ok;
 
 use bytes::BytesMut;
-use std::cmp;
 use std::io;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
 #[tokio::test]
-async fn write_all() {
+async fn write() {
     struct Wr {
         buf: BytesMut,
         cnt: usize,
@@ -23,12 +22,9 @@ async fn write_all() {
             _cx: &mut Context<'_>,
             buf: &[u8],
         ) -> Poll<io::Result<usize>> {
-            let n = cmp::min(4, buf.len());
-            let buf = &buf[0..n];
-
-            self.cnt += 1;
-            self.buf.extend(buf);
-            Ok(buf.len()).into()
+            assert_eq!(self.cnt, 0);
+            self.buf.extend(&buf[0..4]);
+            Ok(4).into()
         }
 
         fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<io::Result<()>> {
@@ -45,7 +41,7 @@ async fn write_all() {
         cnt: 0,
     };
 
-    assert_ok!(wr.write_all(b"hello world").await);
-    assert_eq!(wr.buf, b"hello world"[..]);
-    assert_eq!(wr.cnt, 3);
+    let n = assert_ok!(wr.write(b"hello world").await);
+    assert_eq!(n, 4);
+    assert_eq!(wr.buf, b"hell"[..]);
 }
