+++ tokio-io/tests/write_all.rs
@@ -1,16 +1,17 @@
 #![deny(warnings, rust_2018_idioms)]
 #![feature(async_await)]
 
-use tokio::io::{AsyncWrite, AsyncWriteExt};
+use tokio_io::{AsyncWrite, AsyncWriteExt};
 use tokio_test::assert_ok;
 
 use bytes::BytesMut;
+use std::cmp;
 use std::io;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
 #[tokio::test]
-async fn write() {
+async fn write_all() {
     struct Wr {
         buf: BytesMut,
         cnt: usize,
@@ -22,9 +23,12 @@ async fn write() {
             _cx: &mut Context<'_>,
             buf: &[u8],
         ) -> Poll<io::Result<usize>> {
-            assert_eq!(self.cnt, 0);
-            self.buf.extend(&buf[0..4]);
-            Ok(4).into()
+            let n = cmp::min(4, buf.len());
+            let buf = &buf[0..n];
+
+            self.cnt += 1;
+            self.buf.extend(buf);
+            Ok(buf.len()).into()
         }
 
         fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<io::Result<()>> {
@@ -41,7 +45,7 @@ async fn write() {
         cnt: 0,
     };
 
-    let n = assert_ok!(wr.write(b"hello world").await);
-    assert_eq!(n, 4);
-    assert_eq!(wr.buf, b"hell"[..]);
+    assert_ok!(wr.write_all(b"hello world").await);
+    assert_eq!(wr.buf, b"hello world"[..]);
+    assert_eq!(wr.cnt, 3);
 }
