+++ tokio/src/io.rs
@@ -36,29 +36,12 @@
 //! [`ErrorKind`]: enum.ErrorKind.html
 //! [`Result`]: type.Result.html
 
-mod async_buf_read_ext;
-mod async_read_ext;
-mod async_write_ext;
-mod copy;
-mod flush;
-mod lines;
-mod read;
-mod read_exact;
-mod read_line;
-mod read_to_end;
-mod read_to_string;
-mod read_until;
-mod write;
-mod write_all;
-
-pub use self::async_buf_read_ext::AsyncBufReadExt;
-pub use self::async_read_ext::AsyncReadExt;
-pub use self::async_write_ext::AsyncWriteExt;
-
 // standard input, output, and error
 #[cfg(feature = "fs")]
 pub use tokio_fs::{stderr, stdin, stdout, Stderr, Stdin, Stdout};
-pub use tokio_io::{AsyncBufRead, AsyncRead, AsyncWrite};
+pub use tokio_io::{
+    AsyncBufRead, AsyncBufReadExt, AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt,
+};
 
 // Re-export io::Error so that users don't have to deal
 // with conflicts when `use`ing `tokio::io` and `std::io`.
